Move all mutating functions on SpotsProtocol into a new file
